Init2

+++ src/main/java/web/model/User.java
@@ -1,7 +1,5 @@
 package web.model;
 
-import org.springframework.context.annotation.Configuration;
-
 import javax.persistence.*;
 
 @Entity
@@ -9,7 +7,7 @@ import javax.persistence.*;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "name")
     private String name;
@@ -23,18 +21,18 @@ public class User {
     public User() {
     }
 
-    public User(Long id, String name, String surname, int age) {
+    public User(int id, String name, String surname, int age) {
         this.id = id;
         this.name = name;
         this.surname = surname;
         this.age = age;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
